+++ crates/matrix-sdk/tests/integration/matrix_auth.rs
@@ -4,6 +4,7 @@ use assert_matches::assert_matches;
 use matrix_sdk::{
     config::RequestConfig,
     matrix_auth::{MatrixSession, MatrixSessionTokens},
+    test_utils::{logged_in_client_with_server, no_retry_test_client_with_server},
     AuthApi, AuthSession, Client, RumaApiError,
 };
 use matrix_sdk_base::SessionMeta;
@@ -29,10 +30,6 @@ use url::Url;
 use wiremock::{
     matchers::{method, path},
     Mock, MockServer, Request, ResponseTemplate,
-};
-
-use crate::{
-    logged_in_client_with_server, no_retry_test_client_with_server, test_client_builder_with_server,
 };
 
 #[async_test]
@@ -597,7 +594,7 @@ async fn test_login_doesnt_fail_if_cross_signing_bootstrapping_failed() {
 async fn test_login_with_cross_signing_bootstrapping_already_bootstrapped() {
     // Even if we enabled cross-signing bootstrap for another device, it won't
     // restart the procedure.
-    let (builder, server) = test_client_builder_with_server().await;
+    let (builder, server) = matrix_sdk::test_utils::test_client_builder_with_server().await;
 
     Mock::given(method("POST"))
         .and(path("/_matrix/client/r0/login"))
